+++ crates/ryo3-jiff/src/ry_offset.rs
@@ -115,7 +115,7 @@ impl RyOffset {
     }
 
     #[expect(clippy::wrong_self_convention)]
-    fn to_dict<'py>(&self, py: Python<'py>) -> PyResult<Bound<'py, PyDict>> {
+    pub(crate) fn to_dict<'py>(&self, py: Python<'py>) -> PyResult<Bound<'py, PyDict>> {
         let dict = PyDict::new(py);
         dict.set_item(crate::interns::seconds(py), self.seconds())?;
         dict.set_item(crate::interns::fmt(py), self.py_to_string())?;