+++ crates/ryo3-macro-rules/src/py_errs.rs
@@ -5,26 +5,6 @@
 //! py_*_err! creates a `Result<_, PyErr>`
 
 #[macro_export]
-macro_rules! py_type_error {
-    () => {
-        ::pyo3::exceptions::PyTypeError::new_err("type error")
-    };
-    ($($arg:tt)+) => {
-        ::pyo3::exceptions::PyTypeError::new_err(::std::format!($($arg)+))
-    };
-}
-
-#[macro_export]
-macro_rules! py_type_err {
-    () => {
-        ::std::result::Result::Err(::pyo3::exceptions::PyTypeError::new_err("type error"))
-    };
-    ($($arg:tt)+) => {
-        ::std::result::Result::Err(::pyo3::exceptions::PyTypeError::new_err(::std::format!($($arg)+)))
-    };
-}
-
-#[macro_export]
 macro_rules! py_overflow_error {
     () => {
         ::pyo3::exceptions::PyOverflowError::new_err("overflow error")
@@ -41,6 +21,46 @@ macro_rules! py_overflow_err {
     };
     ($($arg:tt)+) => {
         ::std::result::Result::Err(::pyo3::exceptions::PyOverflowError::new_err(::std::format!($($arg)+)))
+    };
+}
+
+#[macro_export]
+macro_rules! py_runtime_error {
+    () => {
+        ::pyo3::exceptions::PyRuntimeError::new_err("runtime error")
+    };
+    ($($arg:tt)+) => {
+        ::pyo3::exceptions::PyRuntimeError::new_err(::std::format!($($arg)+))
+    };
+}
+
+#[macro_export]
+macro_rules! py_runtime_err {
+    () => {
+        ::std::result::Result::Err(::pyo3::exceptions::PyRuntimeError::new_err("runtime error"))
+    };
+    ($($arg:tt)+) => {
+        ::std::result::Result::Err(::pyo3::exceptions::PyRuntimeError::new_err(::std::format!($($arg)+)))
+    };
+}
+
+#[macro_export]
+macro_rules! py_type_error {
+    () => {
+        ::pyo3::exceptions::PyTypeError::new_err("type error")
+    };
+    ($($arg:tt)+) => {
+        ::pyo3::exceptions::PyTypeError::new_err(::std::format!($($arg)+))
+    };
+}
+
+#[macro_export]
+macro_rules! py_type_err {
+    () => {
+        ::std::result::Result::Err(::pyo3::exceptions::PyTypeError::new_err("type error"))
+    };
+    ($($arg:tt)+) => {
+        ::std::result::Result::Err(::pyo3::exceptions::PyTypeError::new_err(::std::format!($($arg)+)))
     };
 }