+++ tests/std/test_metadata.py
@@ -11,10 +11,8 @@ def test_metadata() -> None:
     assert isinstance(md, ry.Metadata)
     d = md.to_py()
     assert "file_type" in d
-    assert isinstance(d["file_type"], dict)
-    assert isinstance(d["file_type"]["is_dir"], bool)
-    assert isinstance(d["file_type"]["is_file"], bool)
-    assert isinstance(d["file_type"]["is_symlink"], bool)
+    assert isinstance(d["file_type"], str)
+    assert d["file_type"] in ("file", "directory", "symlink")
     assert isinstance(d["len"], int)
     assert isinstance(d["readonly"], bool)
     assert isinstance(d["accessed"], pydt.datetime)